requirements.txt
requests
secure-smtplib
pyyaml
